requirements.txt
flask
openai==1.3.0
google-generativeai>=0.8.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytesseract
pillow
python-dotenv
requests
werkzeug
gunicorn
python-docx
PyPDF2opencv-python-headless
numpy
